refactor : 인수 테스트 단위 테스트 분리
